test : auth, user, friend test added

+++ src/main/java/com/chat/chattingserver/repository/UserRepository.java
@@ -3,11 +3,10 @@ import com.chat.chattingserver.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUserid(String userid);
     Optional<User> findByNickname(String nickname);
